requirements.txt
asarPy==1.0.1
asgiref==3.8.1
Django==5.0.6
django-rest-framework==0.1.0
djangorestframework==3.15.2
psycopg2-binary==2.9.9
sqlparse==0.5.0
